# Backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: SocietyEase API
  description: |
    API specification for the SocietyEase backend (Flask). Authentication uses JWT Bearer tokens. Each API endpoint is mapped to its corresponding user story using the x-user-story extension for traceability between the implementation and the project requirements.

    ## User story mapping

    Every operation carries an `x-user-story` extension. The stories are:

    | ID | User story |
    |---|---|
    | **US-01** | As a **Treasurer**, I want to generate and track maintenance invoices for every flat, so that dues are billed consistently and I can see what is outstanding. |
    | **US-02** | As a **Resident**, I want to raise a complaint about a problem in my flat or the building, so that the society has a record of it and someone is accountable for fixing it. |
    | **US-03** | As a **Resident**, I want to track the status of my complaint from open to closed, so that I know whether it is being worked on without having to chase anyone. |
    | **US-04** | As a **Association manager**, I want to assign a complaint to a maintenance worker, so that the right person is responsible and the job reaches their queue. |
    | **US-05** | As a **Owner**, I want to see the invoices raised against the flat I own, so that I can keep track of what my tenant or I owe the society. |
...

    The same table is maintained in `docs/USER_STORIES.md`.
  version: 4.1.0
servers:
- url: http://localhost:5000
...
paths:
...
  /api/events/:
    get:
      tags:
      - Events
      summary: List events
      description: |-
        Lists active society events/meetings, soonest first, optionally filtered. **User Story — US-20**

        Access: any authenticated, active user.
      x-user-story: US-20
      security:
      - bearerAuth: []
      parameters:
      - name: q
        in: query
        schema:
          type: string
        description: Free-text search over title.
      - name: event_type
        in: query
        schema:
          type: string
          enum:
          - MEETING
          - EVENT
          - HOLIDAY
          - DEADLINE
          - OTHER
      - name: from
        in: query
        schema:
          type: string
        description: Inclusive lower bound on event_date (YYYY-MM-DD).
      - name: to
        in: query
        schema:
          type: string
        description: Inclusive upper bound on event_date (YYYY-MM-DD).
      responses:
        '200':
          description: Active events, soonest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: events_get_events
      x-required-role: any authenticated, active user
    post:
      tags:
      - Events
      summary: Create event
      description: |-
        Creates a society event/meeting. **User Story — US-20**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-20
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
        description: JSON object. Unknown fields are ignored; blank optional values are stored as null.
      responses:
        '201':
          description: Event created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: title or event_date is missing, or event_type is not a valid enum value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: events_add_event
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
  /api/events/{eid}:
    put:
      tags:
      - Events
      summary: Update event
      description: |-
        Updates a society event/meeting. Only supplied fields are changed. **User Story — US-20**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-20
      security:
      - bearerAuth: []
      parameters:
      - name: eid
        in: path
        required: true
        schema:
          type: integer
        description: Identifier of the target record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequest'
      responses:
        '200':
          description: Event updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No event with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: events_update_event
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
    delete:
      tags:
      - Events
      summary: Delete event
      description: |-
        Soft-deletes an event (sets is_active=false), the same pattern used by Announcements. **User Story — US-20**

        Access: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER.
      x-user-story: US-20
      security:
      - bearerAuth: []
      parameters:
      - name: eid
        in: path
        required: true
        schema:
          type: integer
        description: Identifier of the target record.
      responses:
        '200':
          description: Event deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: No event with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: events_delete_event
      x-required-role: ADMIN, SYSTEM_ADMIN, TREASURER or COMMITTEE_MEMBER
  /api/events/upcoming:
    get:
      tags:
      - Events
      summary: Merged upcoming-deadlines feed
      description: |-
        A single chronologically sorted, role-aware feed merging manual events, the caller's own unpaid invoices (or every flat's for an admin), active polls, maintenance tasks (admin and the assigned worker), and equipment service dates (admin only). Calendar (.ics) export is not implemented — this is an in-app feed only. **User Story — US-20**

        Access: any authenticated, active user.
      x-user-story: US-20
      security:
      - bearerAuth: []
      parameters:
      - name: days
        in: query
        schema:
          type: integer
          default: 30
        description: How many days ahead to include. 1-365.
      responses:
        '200':
          description: Upcoming items, soonest (or most overdue) first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UpcomingItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
      operationId: events_upcoming
      x-required-role: any authenticated, active user
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
      - error
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        event_type:
          type: string
          enum:
          - MEETING
          - EVENT
          - HOLIDAY
          - DEADLINE
          - OTHER
        event_date:
          type: string
        event_time:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        created_by:
          type: integer
        created_at:
          type: string
    EventCreateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        event_type:
          type: string
          enum:
          - MEETING
          - EVENT
          - HOLIDAY
          - DEADLINE
          - OTHER
        event_date:
          type: string
        event_time:
          type: string
        location:
          type: string
      required:
      - title
      - event_date
    UpcomingItem:
      type: object
      description: One entry in the merged /api/events/upcoming feed, merged from manual events, unpaid invoices, active polls, maintenance tasks and equipment service dates.
      properties:
        type:
          type: string
          description: 'The event type/category, e.g. MEETING, DEADLINE, MAINTENANCE.'
        title:
          type: string
        date:
          type: string
        days_until:
          type: integer
          description: Negative when the date is already in the past.
        severity:
          type: string
          enum:
          - urgent
          - high
          - low
        source:
          type: string
          enum:
          - event
          - invoice
          - poll
          - maintenance
          - equipment
        link:
          type: string
          description: Frontend route this item corresponds to.
  responses:
    BadRequest:
      description: Validation failed — a required field is missing, or a value has the wrong type/format
        (bad date, non-numeric amount, unknown enum value, or a body that is not a JSON object).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'category must be one of: PLUMBING, ELECTRICAL, CLEANING, SECURITY, OTHER'
    Unauthorized:
      description: 'No JWT was supplied, or it is invalid or malformed. Note: flask-jwt-extended emits
        these under a `msg` key rather than `error`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            msg: Missing Authorization Header
    Forbidden:
      description: 'The caller is authenticated but not allowed: either the account has been deactivated,
        or the user''s role is not permitted to perform this action, or the resource belongs to another
        resident.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: You are not allowed to perform this action
    NotFound:
      description: The requested resource does not exist. Also returned when the JWT is valid but the
        user it identifies has been removed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
...
    ServerError:
      description: Unexpected server error. All errors are returned as JSON; the API never returns an
        HTML error page.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
...